+++ CHANGELOG.md
@@ -8,9 +8,10 @@
 
 ### Fixed
 
--  Use maintenance fixed branch of presage with updated root CA ([#190])
+-  Use maintenance fixed branch of presage with updated root CA ([#189], [#190])
 
 [#182]: https://github.com/boxdot/gurk-rs/pull/182
+[#189]: https://github.com/boxdot/gurk-rs/pull/189
 [#190]: https://github.com/boxdot/gurk-rs/pull/190
 
 ## 0.2.5