+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
-## Unreleased (0.3.0-dev)
+## 0.3.0
+
+⚠️ This release requires relinking.
 
 ### Added